Introduce the Patcher package.
The patcher package exposes a Patcher, which allows us to apply patches
to the server resources in a convenient matter.

It mimics most of the `kubectl apply` functionality and enforces a set
of defaults. The key difference is that Kubekit tracks its own
configuration and leaves the kubectl configuration alone.

When using kubectl, it will annotate the last applied configuration to an object
under the key `kubectl.kubernetes.io/last-applied-configuration`. Kubekit will
do the same, but under its own key, `kubekit/last-applied-configuration`. This
is done so that we don't conflict with manual `kubectl apply` interactions.

By specifying this annotation, we can perform a three way merge patch to ensure
that we only update the fields that we should be updating specified by the CRD
change. We enforced the use of this annotation for two distinct reasons:

- controller restarts; when a controller runs and receives an update, we have
  the original state (last-applied-configuration) available together with the
  modified state. This does not survive between controller restarts when a CRD
  gets created again within the controller. Creation of objects could be ignored
  when the object already exists, but this leads us to the second point;
- three-way merge patches are not error free, especially when there is an
  interaction from an external party. By separating this from the kubectl
  annotation we 1) lower the conflict barrier; 2) allow a mechanism for further
  debugging and restoring state.
- with using an annotation and a patch, we allow other parties to edit the
  objects from outside of kubekit. This way other controllers could potentially
  interact with this as well. By specifying a name to the patcher, we ensure
  that the annotations are unique - to a certain extent - across multiple
  controllers. This allows each controller to perform its own updates
  separately.